[otbnsim] Moved the kmac interface simulation to its own file.
Code is only moved in htis commit, not changed. This is preparation to
logically group the KMAC block with OTBN's top-level state, rather than with
its WSRs, which makes it easier to track e.g. that KMAC runs even when OTBN is
stalled.
